COMFY: Move midiplyr driver to main folder
