+++ engines/comfy/anim_file.cpp
@@ -20,7 +20,7 @@
  */
 
 #include "comfy/comfy.h"
-#include "comfy/midiplyr/midiplyr.h"
+#include "comfy/midiplyr.h"
 
 #include "common/debug.h"
 #include "common/endian.h"
